+++ src/wallet/wallet.h
@@ -922,6 +922,8 @@ public:
     std::set< std::set<CTxDestination> > GetAddressGroupings();
     std::map<CTxDestination, CAmount> GetAddressBalances();
 
+    CAmount GetAccountBalance(const std::string& strAccount, int nMinDepth, const isminefilter& filter, bool fAddLockConf);
+    CAmount GetAccountBalance(CWalletDB& walletdb, const std::string& strAccount, int nMinDepth, const isminefilter& filter, bool fAddLockConf);
     std::set<CTxDestination> GetAccountAddresses(const std::string& strAccount) const;
 
     isminetype IsMine(const CTxIn& txin) const;
@@ -1029,6 +1031,7 @@ public:
 
     /* Wallets parameter interaction */
     static bool ParameterInteraction();
+    bool BackupWallet(const std::string& strDest);
 
     /**
      * HD Wallet Functions
